reafacto : changing component name of login showcase and move logics in hooks
