+++ mobile/src/app/showcase/login.tsx
@@ -3,10 +3,9 @@ import { useForm } from 'react-hook-form';
 import { Form } from '@/components/molecules/form/form';
 import { InputGenerator } from '@/utils/generator/input-generator';
 import { LOGIN_INPUT_CONFIGS } from '@/configs/inputs/login-input.config';
-import { login } from '@/services/auth';
 import { LoginCredentials } from '@/types/login';
-import { useMutation } from '@tanstack/react-query';
 import { Button } from '@/components/atoms/buttons/button';
+import { useAuth } from '@/hooks/useAuth';
 
 export default function ShowcaseLoginScreen() {
   const form = useForm<LoginCredentials>({
@@ -16,18 +15,10 @@ export default function ShowcaseLoginScreen() {
     },
   });
 
-  const loginMutation = useMutation({
-    mutationFn: (data: LoginCredentials) => login(data),
-    onSuccess: (response) => {
-      console.log('Login successful:', response);
-    },
-    onError: (error) => {
-      console.error('Login failed:', error);
-    },
-  });
+  const { login, isLoading } = useAuth();
 
   const handleSubmit = async (data: LoginCredentials) => {
-    loginMutation.mutate(data);
+    login(data);
   };
 
   return (
@@ -43,13 +34,13 @@ export default function ShowcaseLoginScreen() {
         <InputGenerator
           configs={LOGIN_INPUT_CONFIGS}
           defaultValues={form.getValues()}
-          control={form.control} // Add this line to pass the form control
+          control={form.control}
         />
         <Button
           variant="default"
           size="lg"
           onPress={form.handleSubmit(handleSubmit)}
-          disabled={loginMutation.isPending}
+          disabled={isLoading}
         >
           Se connecter
         </Button>
